feat: add license
