+++ pkg/app/server/binding/base_type_decoder.go
@@ -1,10 +1,50 @@
-package binding_v2
+/*
+ * Copyright 2022 CloudWeGo Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * MIT License
+ *
+ * Copyright (c) 2019-present Fenny and Contributors
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ *
+ * This file may have been modified by CloudWeGo authors. All CloudWeGo
+ * Modifications are Copyright 2022 CloudWeGo Authors
+ */
+
+package binding
 
 import (
 	"fmt"
 	"reflect"
 
-	"github.com/cloudwego/hertz/pkg/app/server/binding_v2/text_decoder"
+	"github.com/cloudwego/hertz/pkg/app/server/binding/text_decoder"
 	"github.com/cloudwego/hertz/pkg/common/utils"
 	"github.com/cloudwego/hertz/pkg/protocol"
 )
@@ -26,7 +66,6 @@ func (d *baseTypeFieldTextDecoder) Decode(req *protocol.Request, params PathPara
 	var err error
 	var text string
 	var defaultValue string
-	// 最大努力交付，对齐 hertz 现有设计
 	for _, tagInfo := range d.tagInfos {
 		if tagInfo.Key == jsonTag {
 			continue
@@ -37,7 +76,6 @@ func (d *baseTypeFieldTextDecoder) Decode(req *protocol.Request, params PathPara
 		ret := tagInfo.Getter(req, params, tagInfo.Value)
 		defaultValue = tagInfo.Default
 		if len(ret) != 0 {
-			// 非数组/切片类型，只取第一个值作为只
 			text = ret[0]
 			err = nil
 			break
@@ -56,12 +94,10 @@ func (d *baseTypeFieldTextDecoder) Decode(req *protocol.Request, params PathPara
 		return nil
 	}
 
-	// 得到该field的非nil值
+	// get the non-nil value for the field
 	reqValue = GetFieldValue(reqValue, d.parentIndex)
-	// 根据最终的 Struct，获取对应 field 的 reflect.Value
 	field := reqValue.Field(d.index)
 	if field.Kind() == reflect.Ptr {
-		// 如果是指针则新建一个reflect.Value，然后赋值给指针
 		t := field.Type()
 		var ptrDepth int
 		for t.Kind() == reflect.Ptr {
