+++ pkg/app/server/binding/reflect.go
@@ -1,0 +1,113 @@
+/*
+ * Copyright 2022 CloudWeGo Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * MIT License
+ *
+ * Copyright (c) 2019-present Fenny and Contributors
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ *
+ * This file may have been modified by CloudWeGo authors. All CloudWeGo
+ * Modifications are Copyright 2022 CloudWeGo Authors
+ */
+
+package binding
+
+import (
+	"reflect"
+	"unsafe"
+)
+
+func valueAndTypeID(v interface{}) (reflect.Value, uintptr) {
+	header := (*emptyInterface)(unsafe.Pointer(&v))
+
+	rv := reflect.ValueOf(v)
+	return rv, header.typeID
+}
+
+type emptyInterface struct {
+	typeID  uintptr
+	dataPtr unsafe.Pointer
+}
+
+// ReferenceValue convert T to *T, the ptrDepth is the count of '*'.
+func ReferenceValue(v reflect.Value, ptrDepth int) reflect.Value {
+	switch {
+	case ptrDepth > 0:
+		for ; ptrDepth > 0; ptrDepth-- {
+			vv := reflect.New(v.Type())
+			vv.Elem().Set(v)
+			v = vv
+		}
+	case ptrDepth < 0:
+		for ; ptrDepth < 0 && v.Kind() == reflect.Ptr; ptrDepth++ {
+			v = v.Elem()
+		}
+	}
+	return v
+}
+
+func GetNonNilReferenceValue(v reflect.Value) (reflect.Value, int) {
+	var ptrDepth int
+	t := v.Type()
+	elemKind := t.Kind()
+	for elemKind == reflect.Ptr {
+		t = t.Elem()
+		elemKind = t.Kind()
+		ptrDepth++
+	}
+	val := reflect.New(t).Elem()
+	return val, ptrDepth
+}
+
+func GetFieldValue(reqValue reflect.Value, parentIndex []int) reflect.Value {
+	for _, idx := range parentIndex {
+		if reqValue.Kind() == reflect.Ptr && reqValue.IsNil() {
+			nonNilVal, ptrDepth := GetNonNilReferenceValue(reqValue)
+			reqValue.Set(ReferenceValue(nonNilVal, ptrDepth))
+		}
+		for reqValue.Kind() == reflect.Ptr {
+			reqValue = reqValue.Elem()
+		}
+		reqValue = reqValue.Field(idx)
+	}
+
+	// It is possible that the parent struct is also a pointer,
+	// so need to create a non-nil reflect.Value for it at runtime.
+	for reqValue.Kind() == reflect.Ptr {
+		if reqValue.IsNil() {
+			nonNilVal, ptrDepth := GetNonNilReferenceValue(reqValue)
+			reqValue.Set(ReferenceValue(nonNilVal, ptrDepth))
+		}
+		reqValue = reqValue.Elem()
+	}
+
+	return reqValue
+}
