--- pkg/app/server/binding_v2/binder.go
@@ -1,17 +1,0 @@
-package binding_v2
-
-import (
-	"github.com/cloudwego/hertz/pkg/protocol"
-)
-
-// PathParams parameter acquisition interface on the URL path
-type PathParams interface {
-	Get(name string) (string, bool)
-}
-
-type Binder interface {
-	Name() string
-	Bind(*protocol.Request, PathParams, interface{}) error
-}
-
-var DefaultBinder Binder = &Bind{}
