--- pkg/app/server/binding_v2/default_validator.go
@@ -1,54 +1,0 @@
-package binding_v2
-
-import (
-	"reflect"
-	"sync"
-
-	"github.com/go-playground/validator/v10"
-)
-
-var _ StructValidator = (*defaultValidator)(nil)
-
-type defaultValidator struct {
-	once     sync.Once
-	validate *validator.Validate
-}
-
-// ValidateStruct receives any kind of type, but only performed struct or pointer to struct type.
-func (v *defaultValidator) ValidateStruct(obj interface{}) error {
-	if obj == nil {
-		return nil
-	}
-
-	value := reflect.ValueOf(obj)
-	switch value.Kind() {
-	case reflect.Ptr:
-		return v.ValidateStruct(value.Elem().Interface())
-	case reflect.Struct:
-		return v.validateStruct(obj)
-	default:
-		return nil
-	}
-}
-
-// validateStruct receives struct type
-func (v *defaultValidator) validateStruct(obj interface{}) error {
-	v.lazyinit()
-	return v.validate.Struct(obj)
-}
-
-func (v *defaultValidator) lazyinit() {
-	v.once.Do(func() {
-		v.validate = validator.New()
-		v.validate.SetTagName("validate")
-	})
-}
-
-// Engine returns the underlying validator engine which powers the default
-// Validator instance. This is useful if you want to register custom validations
-// or struct level validations. See validator GoDoc for more info -
-// https://pkg.go.dev/github.com/go-playground/validator/v10
-func (v *defaultValidator) Engine() interface{} {
-	v.lazyinit()
-	return v.validate
-}
