--- pkg/app/server/binding_v2/text_decoder/bool.go
@@ -1,17 +1,0 @@
-package text_decoder
-
-import (
-	"reflect"
-	"strconv"
-)
-
-type boolDecoder struct{}
-
-func (d *boolDecoder) UnmarshalString(s string, fieldValue reflect.Value) error {
-	v, err := strconv.ParseBool(s)
-	if err != nil {
-		return err
-	}
-	fieldValue.SetBool(v)
-	return nil
-}
