--- pkg/app/server/binding_v2/text_decoder/string.go
@@ -1,11 +1,0 @@
-package text_decoder
-
-import "reflect"
-
-type stringDecoder struct{}
-
-func (d *stringDecoder) UnmarshalString(s string, fieldValue reflect.Value) error {
-	// todo: 优化一下
-	fieldValue.SetString(s)
-	return nil
-}
