--- pkg/app/server/binding_v2/text_decoder/unit.go
@@ -1,19 +1,0 @@
-package text_decoder
-
-import (
-	"reflect"
-	"strconv"
-)
-
-type uintDecoder struct {
-	bitSize int
-}
-
-func (d *uintDecoder) UnmarshalString(s string, fieldValue reflect.Value) error {
-	v, err := strconv.ParseUint(s, 10, d.bitSize)
-	if err != nil {
-		return err
-	}
-	fieldValue.SetUint(v)
-	return nil
-}
